[refactor]  Impove 'ProfileModal' Component
- ProfileModal 컴포넌트를 따로 폴더로 빼냄
- 중복되는 interface는 extends 로 변경
- style은 styled.ts로 이동

#9
